+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Set endpoint config URL to `playground.ethanppl.com`
 - Handle error when player concurrently start/end games
-- Fix Super tic tac toe: timer round off is incorrect
+- Fix super tic tac toe: timer round off is incorrect
+- Improvement in super hangman: reset guess letter form after submit
 
 ## [0.2.5] - 2024-12-15